pyproject.toml
[project]
name = "tree-scaper"
version = "0.1.0"
description = "Repository with code for visualizing Forest/Tree structures using Python and PyGame."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logging>=0.4.9.6",
    "pydantic>=2.12.5",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.11",
]

[[tool.mypy.overrides]]
module = ["yaml.*"]
follow_untyped_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]